provide traits for spec2 with nicer names
